pyproject.toml
[tool.poetry]
name = "rin-da-refresher"
version = "0.1.0"
description = "A refresher script for Rin's DeviantArt service"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "rin_da_refresher"}]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
orjson = "^3.8.0"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
cchardet = "^2.1.7"
SQLAlchemy = "^1.4.41"
asyncpg = "^0.26.0"
uvloop = "^0.17.0"
python-dotenv = "^0.21.0"
pysimdjson = "^5.0.2"
rin-exceptions = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
